+++ src/app.rs
@@ -13,7 +13,7 @@ use crate::{
 		AppOption, BlameFilePopup, BranchListPopup, CommitPopup,
 		CompareCommitsPopup, ConfirmPopup, CreateBranchPopup,
 		CreateRemotePopup, ExternalEditorPopup, FetchPopup,
-		FileRevlogPopup, FuzzyFindPopup, HelpPopup,
+		FileRevlogPopup, FuzzyFindPopup, GotoLinePopup, HelpPopup,
 		InspectCommitPopup, LogSearchPopupPopup, MsgPopup,
 		OptionsPopup, PullPopup, PushPopup, PushTagsPopup,
 		RemoteListPopup, RenameBranchPopup, RenameRemotePopup,
@@ -112,6 +112,7 @@ pub struct App {
 	popup_stack: PopupStack,
 	options: SharedOptions,
 	repo_path_text: String,
+	goto_line_popup: GotoLinePopup,
 
 	// "Flags"
 	requires_redraw: Cell<bool>,
@@ -218,6 +219,7 @@ impl App {
 			stashing_tab: Stashing::new(&env),
 			stashlist_tab: StashList::new(&env),
 			files_tab: FilesTab::new(&env),
+			goto_line_popup: GotoLinePopup::new(&env),
 			tab: 0,
 			queue: env.queue,
 			theme: env.theme,
@@ -481,6 +483,7 @@ impl App {
 			msg_popup,
 			confirm_popup,
 			commit_popup,
+			goto_line_popup,
 			blame_file_popup,
 			file_revlog_popup,
 			stashmsg_popup,
@@ -544,7 +547,8 @@ impl App {
 			fetch_popup,
 			options_popup,
 			confirm_popup,
-			msg_popup
+			msg_popup,
+			goto_line_popup
 		]
 	);
 
@@ -904,6 +908,14 @@ impl App {
 			}
 			InternalEvent::CommitSearch(options) => {
 				self.revlog.search(options);
+			}
+			InternalEvent::OpenGotoLinePopup(max_line) => {
+				self.goto_line_popup.open(max_line);
+			}
+			InternalEvent::GotoLine(line) => {
+				if self.blame_file_popup.is_visible() {
+					self.blame_file_popup.goto_line(line);
+				}
 			}
 		}