+++ src/queue.rs
@@ -157,6 +157,10 @@ pub enum InternalEvent {
 	RewordCommit(CommitId),
 	///
 	CommitSearch(LogFilterSearchOptions),
+	///
+	OpenGotoLinePopup(usize),
+	///
+	GotoLine(usize),
 }
 
 /// single threaded simple queue for components to communicate with each other